added redux for some components

+++ frontend/src/account/Signout.js
@@ -2,38 +2,55 @@ import React, { useEffect, useState } from 'react';
 import Container from 'react-bootstrap/esm/Container';
 import { Redirect } from 'react-router';
 import cookie from 'react-cookies';
-import { useDispatch } from 'react-redux';
+import { useDispatch, useSelector } from 'react-redux';
 import { bindActionCreators } from 'redux';
 
 import { get } from '../utils/serverCall';
 import { COOKIE } from '../utils/consts';
-import { actionCreators } from '../reducers/actionCreators';
+import { actionCreators, apiActionCreators } from '../reducers/actionCreators';
 
 function Signout() {
   const dispatch = useDispatch();
-  const { signout, clearCart, clearHomeFilters } = bindActionCreators(
+  const { dispatchSignout, clearCart, clearHomeFilters } = bindActionCreators(
     actionCreators,
     dispatch
   );
   const [logout, setLogout] = useState(false);
+  const { doGet } = bindActionCreators(apiActionCreators, dispatch);
+  const signoutState = useSelector((state) => state.signoutApi);
 
   useEffect(() => {
-    get('/signout')
-      .then(() => {
-        // cookie.remove(COOKIE, { path: '/' });
+    doGet('/signout');
+    // get('/signout')
+    //   .then(() => {
+    //     // cookie.remove(COOKIE, { path: '/' });
+    //     localStorage.clear();
+    //     dispatchSignout();
+    //     clearCart();
+    //     clearHomeFilters();
+    //     setLogout(true);
+    //   })
+    //   .catch(() => {
+    //     console.log('error');
+    //   });
+  }, []);
+
+  useEffect(() => {
+    if (signoutState.status === 1) {
+      if (signoutState.error === '') {
         localStorage.clear();
-        signout();
+        dispatchSignout();
         clearCart();
         clearHomeFilters();
         setLogout(true);
-      })
-      .catch(() => {
+      } else {
         console.log('error');
-      });
-  }, []);
+      }
+    }
+  }, [signoutState]);
 
   if (logout) {
-    return <Redirect to='/signin'></Redirect>;
+    return <Redirect to="/signin" />;
   }
   return <Container> Logging out ...</Container>;
 }
